generations/library_rag/requirements.txt
# Core dependencies
weaviate-client>=4.18.0
flask>=3.0.0
mistralai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
werkzeug>=3.0.0

# MCP Server dependencies
mcp>=1.0.0
pydantic>=2.0.0

# Type checking and static analysis
mypy>=1.8.0
types-Flask>=1.1.0
types-requests>=2.31.0

# Documentation validation
pydocstyle>=6.3.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.23.0
httpx>=0.25.0

# MCP Client (for Python agent applications)
anthropic>=0.39.0  # Claude API
